+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 
 ## [2.7.5] - 2019-07-22
 ### Change
+- gen_libs.list_files:  Refactored function.
 - gen_libs.str_2_list:  Refactored function.
 
 ### Fixed